[Fix] apply code style
