+++ app/src/main/java/io/github/kurramkurram/angerlog/ui/screen/home/HomeScreen.kt
@@ -1,8 +1,11 @@
 package io.github.kurramkurram.angerlog.ui.screen.home
 
 import androidx.activity.compose.LocalActivity
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -11,7 +14,14 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Reviews
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ElevatedCard
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -20,12 +30,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.github.kurramkurram.angerlog.R
+import io.github.kurramkurram.angerlog.ui.AngerLevel
 import io.github.kurramkurram.angerlog.ui.component.dialog.AngerLogBasicDialog
 import io.github.kurramkurram.angerlog.ui.component.layout.AngerLogTabLayout
 import io.github.kurramkurram.angerlog.ui.component.layout.TabContent
@@ -257,6 +270,78 @@ fun HomeScreenLookBackContent(
         ) {
             items(state.lookBackList) { angerLog ->
                 HomeListItem(item = angerLog, onItemClick = { id -> onClick(id) })
+            }
+        }
+    }
+}
+
+/**
+ * ホーム画面のリストの項目.
+ *
+ * @param modifier [Modifier]
+ * @param item リストの表示内容
+ * @param onItemClick リストを選択した時の動作
+ * @param isRoundTop 項目の上に丸みを付けるかどうか
+ * @param isRoundBottom 項目の下に丸みを付けるかどうか
+ */
+@Composable
+fun HomeListItem(
+    modifier: Modifier = Modifier,
+    item: HomeAngerLog,
+    onItemClick: (id: Long) -> Unit,
+    isRoundTop: Boolean = true,
+    isRoundBottom: Boolean = true,
+) {
+    ElevatedCard(
+        modifier =
+            modifier
+                .padding(10.dp)
+                .clickable { onItemClick(item.getId()) },
+        elevation =
+            CardDefaults.cardElevation(
+                defaultElevation = 6.dp,
+            ),
+    ) {
+        Column {
+            val shape =
+                if (isRoundTop && isRoundBottom) {
+                    MaterialTheme.shapes.small
+                } else if (isRoundTop) {
+                    RoundedCornerShape(topStart = 8.dp, topEnd = 8.dp)
+                } else if (isRoundBottom) {
+                    RoundedCornerShape(bottomStart = 8.dp, bottomEnd = 8.dp)
+                } else {
+                    RoundedCornerShape(0.dp)
+                }
+            Row(
+                modifier =
+                    modifier
+                        .fillMaxWidth()
+                        .background(color = MaterialTheme.colorScheme.onPrimary, shape = shape)
+                        .padding(horizontal = 15.dp, vertical = 8.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Column(modifier = modifier.weight(1f)) {
+                    Text(text = item.getDate())
+                    Text(text = item.getEvent(), maxLines = 1, overflow = TextOverflow.Ellipsis)
+                }
+
+                if (item.canShowLookBack()) {
+                    Icon(
+                        modifier = modifier.padding(horizontal = 20.dp),
+                        imageVector = Icons.Outlined.Reviews,
+                        contentDescription = stringResource(R.string.home_review_cd),
+                    )
+                }
+
+                Text(
+                    modifier =
+                        modifier
+                            .clip(CircleShape)
+                            .background(color = AngerLevel().select(level = item.getLevel()).getColor())
+                            .padding(horizontal = 20.dp, vertical = 5.dp),
+                    text = "${item.getLevel()}",
+                )
             }
         }
     }
